+++ flexiapi/app/Space.php
@@ -1,5 +1,21 @@
 <?php
+/*
+    Flexisip Account Manager is a set of tools to manage SIP accounts.
+    Copyright (C) 2020 Belledonne Communications SARL, All rights reserved.
 
+    This program is free software: you can redistribute it and/or modify
+    it under the terms of the GNU Affero General Public License as
+    published by the Free Software Foundation, either version 3 of the
+    License, or (at your option) any later version.
+
+    This program is distributed in the hope that it will be useful,
+    but WITHOUT ANY WARRANTY; without even the implied warranty of
+    MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
+    GNU Affero General Public License for more details.
+
+    You should have received a copy of the GNU Affero General Public License
+    along with this program.  If not, see <http://www.gnu.org/licenses/>.
+*/
 namespace App;
 
 use Illuminate\Database\Eloquent\Factories\HasFactory;
@@ -10,6 +26,8 @@ use Carbon\Carbon;
 class Space extends Model
 {
     use HasFactory;
+
+    protected $with = ['emailServer'];
 
     public const FORBIDDEN_KEYS = [
         'disable_chat_feature',
@@ -58,6 +76,11 @@ class Space extends Model
     public function admins()
     {
         return $this->accounts()->where('admin', true);
+    }
+
+    public function emailServer()
+    {
+        return $this->hasOne(SpaceEmailServer::class);
     }
 
     public function scopeNotFull(Builder $query)