+++ flexiapi/resources/views/admin/space/tabs.blade.php
@@ -5,8 +5,10 @@
         $items[route('admin.spaces.show', $space->id)] = __('Information');
         $items[route('admin.spaces.administration', $space->id)] = __('Administration');
         $items[route('admin.spaces.edit', $space->id)] = __('App Configuration');
+        $items[route('admin.spaces.integration', $space->id)] = __('Integration');
     } else if (auth()->user()->admin) {
         $items[route('admin.spaces.me')] = __('Information');
+        $items[route('admin.spaces.integration', $space->id)] = __('Integration');
     }
 
     $items[route('admin.spaces.configuration', $space->id)] = __('Configuration');