+++ flexiapi/resources/views/provisioning/documentation.blade.php
@@ -1,8 +1,8 @@
 @extends('layouts.main', ['welcome' => true])
 
 @section('content')
-    <div>
-        {{-- This view is only a wrapper for the markdown page --}}
-        {!! $documentation !!}
-    </div>
+<section class="documentation">
+    {{-- This view is only a wrapper for the markdown page --}}
+    {!! $documentation !!}
+</section>
 @endsection